fix test

+++ account-service/src/main/resources/db/changelog/changeset/creat-table-account.sql
@@ -1,6 +1,7 @@
 create table if not exists account
 (
-    id            bigint primary key not null,
+    id            serial
+        primary key not null,
     name          varchar(255)       not null,
     email         varchar(255)       not null,
     phone_number  varchar(255)       not null,
